# Drop mutext at `txnTime`